Fix calculation of start and end times of reservations
The utc_start and utc_end datetime variables need to include the
date (i.e. not just HH:mm) so that the subsequent timezone specific
time can be calculated with the correct offset for datetimes that
occur during daylight savings periods. Otherwise, the time is
calculated according to the current offset which may not be correct.

Additionally, fix the calculation for the end date. Update to always
set the end date by incrementing the user selected end date by one
day. Fixes a bug where reservations are created one day shorter than
expected and users are unable to create reservations for a single
day.

Change-Id: Ib83af6318a5980d8058f766176f94dc0881ac6f0